requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-jalali-date==0.3.1
django-taggit==1.3.0
django-widget-tweaks==1.4.8
idna==2.10
importlib-metadata==1.7.0
jdatetime==3.6.2
Markdown==3.2.2
polib==1.1.0
psycopg2==2.8.5
psycopg2-binary==2.8.4
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
zipp==3.1.0
